Move Control Panel controllers
